Added the separate files for the subscription structure and convertible protocol with its default implementation
